refactor signup form to page pattern
