+++ app/constants/constants.js
@@ -14,6 +14,8 @@ export const MEMBER_INFO_FETCH_DATA_SUCCESS = 'MEMBER_INFO_FETCH_DATA_SUCCESS'
 export const MEMBER_INFO_IS_LOADING = 'MEMBER_INFO_IS_LOADING'
 export const LOAD_COMPLETE_SIGNUP_LOADING = 'LOAD_COMPLETE_SIGNUP_LOADING'
 export const LOAD_COMPLETE_SIGNUP_SUCCESS = 'LOAD_COMPLETE_SIGNUP_SUCCESS'
+export const SIGNUP_SUCCESS = 'SIGNUP_SUCCESS'
+export const LOAD_SIGNUP_SUCCESS = 'LOAD_SIGNUP_SUCCESS'
 export const LOAD_COMPLETE_SIGNUP_FAILURE = 'LOAD_COMPLETE_SIGNUP_FAILURE'
 export const CREATE_USER_LOADING = 'CREATE_USER_LOADING'
 export const CREATE_USER_SUCCESS = 'CREATE_USER_SUCCESS'
