+++ app/routes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { withRouter, Switch, Route } from 'react-router-dom'
-import SignupContainer from './components/SignupForm/SignupContainer'
+import SignupPage from './pages/SignupPage'
 import CreateUserPage from './pages/CreateUserPage'
 import BriefResponse from './components/BriefResponse/BriefResponse'
 import ResetPasswordRouter from './components/ResetPassword/ResetPasswordRoutes'
@@ -17,8 +17,8 @@ export default withRouter(RootContainer)
 
 export const Routes = () =>
   <Switch>
-    <Route exact path={rootPath} component={SignupContainer} />
-    <Route path={`${rootPath}/signup`} component={SignupContainer} />
+    <Route exact path={rootPath} component={SignupPage} />
+    <Route path={`${rootPath}/signup`} component={SignupPage} />
     <Route path={`${rootPath}/createuser/:tokenstring`} component={CreateUserPage} />
     <Route path={`${rootPath}/brief/:brief_id/respond`} component={BriefResponse} />
     <Route path={`${rootPath}/reset-password`} component={ResetPasswordRouter} />
